added apt-cacher-host and mirror arguments

+++ reconstructor/distro/debian.py
@@ -2,18 +2,35 @@
 import os
 from base import BaseDistro
 import logging
+import tarfile
 from subprocess import Popen, PIPE
+try:
+    import simplejson as json
+except ImportError:
+    import json
 
 class Debian(BaseDistro):
-    def __init__(self, version='squeeze', arch='i386', username='debian', packages=[],\
-        name='DebianCustom'):
+    def __init__(self, project=None, version='squeeze', arch='i386', username='debian', packages=[],\
+        name='DebianCustom', apt_cacher_host=None, mirror=None):
         super(Debian, self).__init__()
         self._log = logging.getLogger('debian')
+        self._project = project
         self._name = name
         self._version = version
         self._arch = arch
         self._username = username
         self._packages = packages
+        if apt_cacher_host and apt_cacher_host != '':
+            self._apt_cacher_host = apt_cacher_host
+        else:
+            self._apt_cacher_host = None
+        if mirror and mirror != '':
+            self._mirror = mirror
+        else:
+            self._mirror = 'http://ftp.us.debian.org/debian'
+        # load project if needed
+        if self._project:
+            self._load_project()
     
     @property
     def name(self):
@@ -35,16 +52,39 @@ class Debian(BaseDistro):
     def packages(self):
         return self._packages
 
+    def _load_project(self):
+        if not os.path.exists(self._project):
+            self._log.error('Unable to find project file {0}'.format(self._project))
+            return None
+        self._log.debug('Extracting project {0}'.format(self._project))
+        tf = tarfile.open(self._project)
+        tf.extractall(self._build_dir)
+        prj = json.loads(open(os.path.join(self._build_dir, 'project.json'), 'r').read())
+        self._name = prj['name']
+        self._arch = prj['arch']
+        self._version = prj['distro']['codename']
+        self._packages = [x['name'] for x in prj['packages']]
+
     def build(self):
         self._log.debug('build')
         cmd = "cd {0} && ".format(self._build_dir)
-        cmd += "lb config --debian-installer true --debian-installer-gui true "
+        cmd += "lb config "
         cmd += "-a {0} ".format(self._arch)
         cmd += "-d {0} ".format(self._version)
+        cmd += "--debian-installer live "
+        cmd += "--debian-installer-gui true "
         cmd += "--iso-application \"{0}\" ".format(self._copyright)
         cmd += "--iso-volume \"{0}\" ".format(self._name)
+        cmd += "--archive-areas \"main contrib non-free\" "
+        #cmd += "--debian-installer-distribution {0} ".format(self._version)
+        cmd += "--debian-installer-distribution daily "
         cmd += "--mode debian "
-        cmd += "-m http://localhost:3142/apt-cacher/ftp.us.debian.org/debian "
+        if self._apt_cacher_host:
+            if self._mirror.find('//') > -1:
+                self._mirror = self._mirror.split('//')[-1]
+            cmd += "-m http://{0}/apt-cacher/{1} ".format(self._apt_cacher_host, self._mirror)
+        else:
+            cmd += "-m {0} ".format(self._mirror)
         cmd += "--username {0} ".format(self._username)
         cmd += "--packages \"{0}\" ".format(' '.join(self._packages))
         self._log.debug(cmd)
@@ -53,9 +93,15 @@ class Debian(BaseDistro):
         cmd = "cd {0} && ".format(self._build_dir)
         cmd += "lb build"
         self._run_command(cmd)
-        iso_file = os.path.join(self._build_dir, 'binary.iso')
-        if os.path.exists(iso_file):
+        if os.path.exists(os.path.join(self._build_dir, 'binary.iso')):
+            iso_file = os.path.join(self._build_dir, 'binary.iso')
+        elif os.path.exists(os.path.join(self._build_dir, 'binary-amd64.iso')):
+            iso_file = os.path.join(self._build_dir, 'binary-amd64.iso')
+        elif os.path.exists(os.path.join(self._build_dir, 'binary-hybrid.iso')):
+            iso_file = os.path.join(self._build_dir, 'binary-hybrid.iso')
+        else:
+            iso_file = None
+        if iso_file and os.path.exists(iso_file):
             os.rename(iso_file, os.path.join(os.getcwd(), '{0}-{1}.iso'.format(self._name, self._arch)))
-
